# api/server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/login": {
      "post": {
        "operationId": "MainController_login",
        "summary": "用户登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/logout": {
      "post": {
        "operationId": "MainController_logout",
        "summary": "退出登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/register": {
      "post": {
        "operationId": "MainController_register",
        "summary": "用户注册",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/registerUser": {
      "get": {
        "operationId": "MainController_registerUser",
        "summary": "账号自助-是否开启用户注册功能",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/captchaImage": {
      "get": {
        "operationId": "MainController_captchaImage",
        "summary": "获取验证图片",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/getInfo": {
      "get": {
        "operationId": "MainController_getInfo",
        "summary": "用户信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/getRouters": {
      "get": {
        "operationId": "MainController_getRouters",
        "summary": "路由信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload": {
      "post": {
        "operationId": "UploadController_singleFileUpload",
        "summary": "文件上传",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/chunk/uploadId": {
      "get": {
        "operationId": "UploadController_getChunkUploadId",
        "summary": "获取切片上传任务Id",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/chunk": {
      "post": {
        "operationId": "UploadController_chunkFileUpload",
        "summary": "文件切片上传",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/chunk/merge": {
      "post": {
        "operationId": "UploadController_chunkMergeFile",
        "summary": "合并切片",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkMergeFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/chunk/result": {
      "get": {
        "operationId": "UploadController_getChunkUploadResult",
        "summary": "获取切片上传结果",
        "parameters": [
          {
            "name": "uploadId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/cos/authorization": {
      "get": {
        "operationId": "UploadController_getAuthorization",
        "summary": "获取cos上传密钥",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "password",
          "uuid"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "nickName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "password",
          "uuid"
        ]
      },
      "FileUploadDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "file"
        ]
      },
      "ChunkMergeFileDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          }
        },
        "required": [
          "uploadId",
          "fileName"
        ]
      },
...
    }
  }
}